Commit #41 - 09.08.2023
Added UserController in the Admin area. Current functionalities are listing all users and allowing overseers to be assigned to different academies.

Added new views and modified existing ones.

Modified HomeworkService to allow access to a overseer's homeworks even if the overseer is currently not assigned to the academy in which the homework was submitted.